Update hybrid GNN-RNN training configuration and attention layers for improved performance
- Enabled mixed precision and XLA compilation in the training configuration to enhance training speed and reduce memory usage.
- Modified attention layer implementations to use dynamic data types for score calculations and masks, improving compatibility and performance across different tensor types.

+++ src/rade_ml/models/hybrid_gnn_rnn/layers/attention_layer.py
@@ -268,10 +268,10 @@ class TargetAttentionLayer(tf.keras.layers.Layer):
         """
         # 1) scaled dot-product
         scores = tf.matmul(q, k, transpose_b=True)
-        scores /= tf.math.sqrt(tf.cast(self.head_units, tf.float32))
+        scores /= tf.math.sqrt(tf.cast(self.head_units, scores.dtype))
 
         # 2) build mask from adjacency (already dense binary after _extract_target_submatrix)
-        mask = tf.cast(adjacency > 0, tf.float32)
+        mask = tf.cast(adjacency > 0, scores.dtype)
         mask = tf.reshape(mask, [1, 1, num_trades, num_trades])  # [1,1,T,T]
 
         # --- masked softmax ---
